fix: more bugs

+++ app/src/main/java/com/weatherpro/models/currentApi/Current.java
@@ -42,7 +42,7 @@ public class Current {
     private int visibility;
     @SerializedName("wind_speed")
     @Expose
-    private int windSpeed;
+    private double windSpeed;
     @SerializedName("wind_deg")
     @Expose
     private int windDeg;
@@ -138,11 +138,11 @@ public class Current {
         this.visibility = visibility;
     }
 
-    public int getWindSpeed() {
+    public double getWindSpeed() {
         return windSpeed;
     }
 
-    public void setWindSpeed(int windSpeed) {
+    public void setWindSpeed(double windSpeed) {
         this.windSpeed = windSpeed;
     }
